+++ accounts-services/src/main/java/com/siliconvalley/accountsservices/dto/inputDtos/PostInputRequestDto.java
@@ -11,6 +11,7 @@ import jakarta.validation.constraints.*;
 import lombok.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -83,16 +84,16 @@ public class PostInputRequestDto {
     private AllConstantHelpers.RELATION bloodRelation;
     private int benAge;
     private AllConstantHelpers.BanksSupported benBank;
-    private long balance;
+    private BigDecimal balance;
     private AllConstantHelpers.UpdateRequest updateRequest;
     private String branchCode;
-    private long transferLimitPerDay;
+    private BigDecimal transferLimitPerDay;
     private int creditScore;
     private AllConstantHelpers.AccountStatus accountStatus;
-    private long approvedLoanLimitBasedOnCreditScore;
+    private BigDecimal approvedLoanLimitBasedOnCreditScore;
     private Boolean anyActiveLoans;
-    private long totLoanIssuedSoFar;
-    private long totalOutStandingAmountPayableToBank;
+    private BigDecimal totLoanIssuedSoFar;
+    private BigDecimal totalOutStandingAmountPayableToBank;
     private MultipartFile customerImage;
     private List<BeneficiaryDto> listOfBeneficiary = new ArrayList<>();
     private List<TransactionsDto> listOfTransactions = new ArrayList<>();
