+++ accounts-services/src/main/java/com/siliconvalley/accountsservices/dto/inputDtos/PutInputRequestDto.java
@@ -9,6 +9,7 @@ import lombok.*;
 import org.springframework.web.multipart.MultipartFile;
 
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -53,18 +54,18 @@ public class PutInputRequestDto {
     private String beneficiaryEmail;
     private String benVoterId;
     private String benDrivingLicense;
-    private long balance;
+    private BigDecimal balance;
     private AllConstantHelpers.UpdateRequest updateRequest;
     private String branchCode;
-    private long transferLimitPerDay;
+    private BigDecimal transferLimitPerDay;
     private int creditScore;
     private String address;
     private String imageName;
     private AllConstantHelpers.AccountStatus accountStatus;
-    private long approvedLoanLimitBasedOnCreditScore;
+    private BigDecimal approvedLoanLimitBasedOnCreditScore;
     private Boolean anyActiveLoans;
-    private long totLoanIssuedSoFar;
-    private long totalOutStandingAmountPayableToBank;
+    private BigDecimal totLoanIssuedSoFar;
+    private BigDecimal totalOutStandingAmountPayableToBank;
     private MultipartFile customerImage;
     private List<Beneficiary> listOfBeneficiary = new ArrayList<>();
     private List<Transactions> listOfTransactions = new ArrayList<>();
